refactor: extract into dedicated crate
