+++ Cargo.toml
@@ -41,6 +41,7 @@ mollusk-svm-fuzz-fs = { path = "fuzz/fs", version = "0.13.1" }
 mollusk-svm-programs-memo = { path = "programs/memo", version = "0.13.1" }
 mollusk-svm-result = { path = "result", version = "0.13.1" }
 mollusk-svm-programs-token = { path = "programs/token", version = "0.13.1" }
+mollusk-svm-programs-token-2022 = { path = "programs/token-2022", version = "0.13.1" }
 num-format = "0.4.4"
 openssl = "0.10.78"
 prost = "0.14"
