+++ programs/token/Cargo.toml
@@ -13,11 +13,7 @@ version = { workspace = true }
 default = ["token", "associated-token", "token-2022"]
 token = []
 associated-token = ["dep:spl-associated-token-account-interface"]
-token-2022 = [
-    "dep:spl-token-2022-interface",
-    "dep:spl-token-group-interface",
-    "dep:spl-token-metadata-interface",
-]
+token-2022 = []
 
 [dependencies]
 mollusk-svm = { workspace = true }
@@ -27,9 +23,3 @@ solana-pubkey = { workspace = true }
 solana-rent = { workspace = true }
 spl-token-interface = { workspace = true }
 spl-associated-token-account-interface = { workspace = true, optional = true  }
-spl-token-2022-interface = { workspace = true, optional = true }
-spl-token-group-interface = { workspace = true, optional = true }
-spl-token-metadata-interface = { workspace = true, optional = true }
-
-[dev-dependencies]
-solana-program-option = { workspace = true }
